BD-2807 / BD-2806: - Code review: Move singNFT to the fio actions and remove types.d.ts and actions.ts also remove email selector.
