--- client/src/redux/nftSignatures/types.d.ts
@@ -1,1 +1,0 @@
-export type ReduxState = ReturnType<typeof store.getState>;
